# backend/openapi.yaml
openapi: 3.0.0
info:
  title: Gram Pragati Setu API
  version: 1.0.0
  description: Backend API for Gram Pragati Setu web portal
servers:
  - url: http://localhost:5000/api
    description: Development server
paths:
...
  /villages:
    get:
      summary: List villages
      tags: [Villages]
      security:
        - bearerAuth: []
      parameters:
        - name: state
          in: query
          schema:
            type: string
        - name: district
          in: query
          schema:
            type: string
        - name: adarsh_candidate
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: List of villages
    post:
      summary: Create village (Officer only)
      tags: [Villages]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, state, district]
              properties:
                name:
                  type: string
                state:
                  type: string
                district:
                  type: string
                block:
                  type: string
                population:
                  type: integer
                latitude:
                  type: number
                longitude:
                  type: number
                geometry:
                  type: object
                baseline_metrics:
                  type: object
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT